Split subreddit.py into a subreddit/ package
Reorganize the 3,700-line praw/models/reddit/subreddit.py into a
praw.models.reddit.subreddit package, with the Subreddit class and each of
its helper classes (Modmail, SubredditFilters, SubredditFlair,
SubredditModeration, SubredditQuarantine, SubredditRelationship,
SubredditStream, SubredditStylesheet, SubredditWiki, and their related
classes) moved to dedicated modules. All classes remain importable from
praw.models.reddit.subreddit and praw.models for backwards compatibility.
